requirements.txt
langchain
openai
chromadb
python-dotenv
# PyPDF2
pypdf
langchain-community
tiktoken
